Added drivers socket status management
